Split out event parsing methods to new module
